Broke out the cli into an includable module so you can write your own wrapper
